Add a scene hot reloader to quickly debug issues
This commit adds an other test program that simply reloads the wanted scene file if it was changed so that you can test changing things in the scene without rebuilding and serializing the scene all the time.
This also adds a parameter to enable saving the Next Controller Input Display scene to a file.